Split dream out into separate files.
