+++ src/inference/dream/gradients.ad.js
@@ -1,0 +1,2 @@
+'use strict';
+'use ad';
