+++ src/inference/dream/sample.ad.js
@@ -2,47 +2,12 @@
 'use ad';
 
 var _ = require('lodash');
-var util = require('../util');
-var paramStruct = require('../params/struct');
-var Trace = require('../trace');
-var guide = require('../guide');
-
-// This estimator makes the following assumptions:
-
-// 1. The model contains exactly one `mapData`.
-
-// 2. Either a) each element of the data array is observed (passed to
-// `observe` as its second argument) exacly once in the corresponding
-// observation function, or b) each element of the data array is
-// itself an array, of which each element is observed exaclty once,
-// and in the order in which they appear in the array. See below for
-// examples of each of these.
-
-// var model = function() {
-//   mapData({data: [x, y]}, function(datum) {
-//     // latent random choices
-//     observe(dist, datum);
-//   });
-// };
-
-// var model = function() {
-//   mapData({data: [[x1, y1], [x2, y2]]}, function(arr) {
-//     // latent random choices
-//     observe(dist, arr[0]);
-//     observe(dist, arr[1]);
-//   });
-// };
-
-
-// I assume that we have one of these schemes, that is data doesn't
-// contain a mixture of both.
+var Trace = require('../../trace');
+var guide = require('../../guide');
 
 module.exports = function(env) {
 
-  // --------------------------------------------------
-  // Coroutine to hallucinate data.
-  // --------------------------------------------------
-  function sampleFantasyCoroutine(wpplFn, s, a, cont) {
+  function dreamSample(wpplFn, s, a, cont) {
     this.wpplFn = wpplFn;
     this.s = s;
     this.a = a;
@@ -58,7 +23,7 @@ module.exports = function(env) {
     env.coroutine = this;
   }
 
-  sampleFantasyCoroutine.prototype = {
+  dreamSample.prototype = {
 
     run: function() {
       return this.wpplFn(_.clone(this.s), function(s, val) {
@@ -154,54 +119,10 @@ module.exports = function(env) {
 
   };
 
-  function sampleFantasy() {
-    var coroutine = Object.create(sampleFantasyCoroutine.prototype);
-    sampleFantasyCoroutine.apply(coroutine, arguments);
+  return function() {
+    var coroutine = Object.create(dreamSample.prototype);
+    dreamSample.apply(coroutine, arguments);
     return coroutine.run();
-  }
-
-  // --------------------------------------------------
-  // Coroutine to estimate gradients.
-  // --------------------------------------------------
-
-
-
-  // --------------------------------------------------
-  // Estimator for use with Optimize.
-  // --------------------------------------------------
-  return function(wpplFn, s, a, options, state, step, cont) {
-    var opts = util.mergeDefaults(options, {
-      samples: 1
-    });
-
-    var objectiveVal = 0;
-    var grad = {};
-
-    return util.cpsLoop(
-      opts.samples,
-      // Loop body.
-      function(i, next) {
-
-        return sampleFantasy(wpplFn, s, a, function(record) {
-
-          console.log(record);
-          return next();
-
-          //return estimateGradient(function(g, objectiveVal_i) {
-            //paramStruct.addEq(grad, g);
-            //objectiveVal += objectiveVal_i;
-            //return next();
-          //});
-
-        });
-
-      },
-      // Continuation.
-      function() {
-        // TODO: divide by num samples.
-        return cont(grad, objectiveVal);
-
-      });
   };
 
 };
