+++ src/inference/optimize.js
@@ -18,7 +18,7 @@ module.exports = function(env) {
   var estimators = {
     ELBO: require('./elbo')(env),
     EUBO: require('./eubo')(env),
-    dream: require('./dream')(env)
+    dream: require('./dream/estimator')(env)
   };
 
   function Optimize(s, k, a, fnOrOptions, maybeOptions) {
